Renamed static symbol to show first in library to get preview
Diancheng mentioned it was difficult to see what the preview of some of the buttons looked like because the static, or default, state appeared at the bottom of the list. This has now been fixed and designers will get a preview of the default or static state first so they can identify which button to use faster.